+++ app/draw/draw-image.js
@@ -1,9 +1,17 @@
-function DrawLine(Visual, obj, options) {
+function DrawImage(Visual, obj, options) {
     const ctx = Visual.ctx;
 
     ctx.beginPath();
     ctx.save();
-    ctx.arc(obj.center[0], obj.center[1], obj.radius, 0, Math.PI * 2);
+
+    const imgcenter = obj.center;
+    const image = obj.image;
+    const sx = -(obj.width / 2);
+    const sy = -(obj.height / 2);
+    const angel = (-obj.rotate * Math.PI / 180) || 0;
+    ctx.translate(imgcenter[0], imgcenter[1]);
+    ctx.rotate(angel);
+    ctx.drawImage(image, sx, sy, obj.width, obj.height);
     ctx.fill();
     if (obj.options.border) {
         ctx.stroke();
@@ -25,19 +33,29 @@ function DrawLine(Visual, obj, options) {
     // const fillStyle = options.fillStyle || '#f00';
     // active
     if (obj.isActive) {
+        let width = obj.width;
+        let height = obj.height;
         ctx.canvas.style.cursor = 'pointer';
-        ctx.lineWidth = 1;
+        ctx.lineWidth = 2;
+
         ctx.beginPath();
+        ctx.save();
+        ctx.translate(imgcenter[0], imgcenter[1]);
+        ctx.rotate(angel);
+
         ctx.strokeStyle = strokeStyle;
-        ctx.moveTo(obj.center[0] - 4, obj.center[1] - 4);
-        ctx.rect(obj.center[0] - 4, obj.center[1] - 4, 8, 8);
+
+        ctx.moveTo(-width / 2 - 2, -height / 2 - 2);
+        ctx.rect(-width / 2 - 2, -height / 2 - 2, width + 4, height + 4);
         ctx.stroke();
         ctx.beginPath();
         ctx.strokeStyle = '#333';
-        ctx.moveTo(obj.center[0] - 3, obj.center[1] - 3);
-        ctx.rect(obj.center[0] - 3, obj.center[1] - 3, 6, 6);
+        ctx.moveTo(-width / 2, -height / 2);
+        ctx.rect(-width / 2, -height / 2, width, height);
         ctx.stroke();
+
+        ctx.restore();
     }
 }
 
-export default DrawLine;
+export default DrawImage;